requirement.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
fast-histogram==0.14
flatbuffers==24.3.25
gast==0.6.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
humanfriendly==10.0
idna==3.10
keras==2.14.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml-dtypes==0.2.0
mpmath==1.3.0
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.14.1
onnxoptimizer==0.3.8
onnxruntime==1.16.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==3.20.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pyclipper==1.3.0.post6
requests==2.32.3
requests-oauthlib==2.0.0
rknn-toolkit2 @ file:///usr/projects/explore/rknn-toolkit2/rknn-toolkit2/packages/rknn_toolkit2-1.6.0%2B81f21f4d-cp311-cp311-linux_x86_64.whl#sha256=f1542483bcb53e99e5bc91d58229b4576046cf3e213522763ad81ef5e202d3d6
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sympy==1.13.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
torch==1.13.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
wrapt==1.14.1
